+++ swagger_parser/lib/src/generator/templates/dart_enum_dto_template.dart
@@ -1,14 +1,61 @@
 import 'package:collection/collection.dart';
 
+import '../../../swagger_parser.dart';
 import '../../utils/case_utils.dart';
 import '../../utils/type_utils.dart';
 import '../../utils/utils.dart';
+import '../models/json_serializer.dart';
 import '../models/universal_data_class.dart';
+import 'dart_import_dto_template.dart';
 
 /// Provides template for generating dart enum DTO
 String dartEnumDtoTemplate(
   UniversalEnumClass enumClass, {
-  required bool freezed,
+  required JsonSerializer jsonSerializer,
+  required bool enumsToJson,
+  required bool unknownEnumValue,
+  required bool markFileAsGenerated,
+}) {
+  if (jsonSerializer == JsonSerializer.dartMappable) {
+    return _dartEnumDartMappableTemplate(
+      enumClass,
+      enumsToJson: enumsToJson,
+      unknownEnumValue: unknownEnumValue,
+      markFileAsGenerated: markFileAsGenerated,
+    );
+  } else {
+    final className = enumClass.name.toPascal;
+    final jsonParam = unknownEnumValue || enumsToJson;
+
+    final values = '${enumClass.items.mapIndexed(
+          (i, e) => _enumValue(
+            i,
+            enumClass.type,
+            e,
+            jsonParam: jsonParam,
+          ),
+        ).join(',\n')}${unknownEnumValue ? ',' : ';'}';
+    final unkownEnumValueStr = unknownEnumValue ? _unkownEnumValue() : '';
+    final constructorStr = jsonParam ? _constructor(className) : '';
+    final fromJsonStr = unknownEnumValue ? _fromJson(className, enumClass) : '';
+    final jsonFieldStr = jsonParam ? _jsonField(enumClass) : '';
+    final toJsonStr = enumsToJson ? _toJson(enumClass, className) : '';
+
+    return '''
+${generatedFileComment(
+      markFileAsGenerated: markFileAsGenerated,
+    )}${dartImportDtoTemplate(jsonSerializer)}
+
+${descriptionComment(enumClass.description)}@JsonEnum()
+enum $className {
+$values$unkownEnumValueStr$constructorStr$fromJsonStr$jsonFieldStr$toJsonStr
+}
+''';
+  }
+}
+
+String _dartEnumDartMappableTemplate(
+  UniversalEnumClass enumClass, {
   required bool enumsToJson,
   required bool unknownEnumValue,
   required bool markFileAsGenerated,
@@ -17,21 +64,18 @@ String dartEnumDtoTemplate(
   final jsonParam = unknownEnumValue || enumsToJson;
 
   final values =
-      '${enumClass.items.mapIndexed((i, e) => _enumValue(i, enumClass.type, e, jsonParam: jsonParam)).join(',\n')}${unknownEnumValue ? ',' : ';'}';
-  final unkownEnumValueStr = unknownEnumValue ? _unkownEnumValue() : '';
-  final constructorStr = jsonParam ? _constructor(className) : '';
-  final fromJsonStr = unknownEnumValue ? _fromJson(className, enumClass) : '';
-  final jsonFieldStr = jsonParam ? _jsonField(enumClass) : '';
-  final toJsonStr = enumsToJson ? _toJson(enumClass, className) : '';
+      '${enumClass.items.mapIndexed((i, e) => _enumValueDartMappable(i, enumClass.type, e, jsonParam: jsonParam)).join(',\n')}${unknownEnumValue ? ',' : ';'}';
 
   return '''
 ${generatedFileComment(
     markFileAsGenerated: markFileAsGenerated,
-  )}import '${freezed ? 'package:freezed_annotation/freezed_annotation.dart' : 'package:json_annotation/json_annotation.dart'}';
+  )}${dartImportDtoTemplate(JsonSerializer.dartMappable)}
 
-${descriptionComment(enumClass.description)}@JsonEnum()
+part '${enumClass.name.toSnake}.mapper.dart';
+
+${descriptionComment(enumClass.description)}@MappableEnum()
 enum $className {
-$values$unkownEnumValueStr$constructorStr$fromJsonStr$jsonFieldStr$toJsonStr
+$values
 }
 ''';
 }
@@ -64,6 +108,16 @@ String _enumValue(
     '''
 ${index != 0 ? '\n' : ''}${descriptionComment(item.description, tab: '  ')}  @JsonValue(${type == 'string' ? "'${item.jsonKey}'" : item.jsonKey})
   ${item.name.toCamel}${jsonParam ? '(${type == 'string' ? "'${item.jsonKey}'" : item.jsonKey})' : ''}''';
+
+String _enumValueDartMappable(
+  int index,
+  String type,
+  UniversalEnumItem item, {
+  required bool jsonParam,
+}) =>
+    '''
+${index != 0 ? '\n' : ''}${descriptionComment(item.description, tab: '  ')}${indentation(2)}@MappableValue(${type == 'string' ? "'${item.jsonKey}'" : item.jsonKey}) 
+${indentation(2)}${item.name.toCamel}''';
 
 String _toJson(UniversalEnumClass enumClass, String className) =>
     '\n\n  ${enumClass.type.toDartType()}? toJson() => json;';