+++ swagger_parser/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.13.0
+- Added support for (`dart_mappable`)[https://github.com/schultek/dart_mappable] serializer
+- Changed `freezed` schema property to `jsonSerializer`, which can be set to `freezed`, `dart_mappable` or `json_serializable` (default).
+- Fixed enum generation name that are defined inside an array
+
 ## 1.12.2
 - Fixes enum duplicate names (#140)[https://github.com/Carapacik/swagger_parser/issues/140]